Update jeu a cliquer.py

+++ appli ica/jeu a cliquer/jeu a cliquer.py
@@ -2,12 +2,12 @@
 import cytron
 
 global nb
-nb = 0
+nb = 0.0
 
 cytron.cy_mkdir("/cytron/sys/app/", "sauvegardes_jeu_a_cliquer")
 
 try:
-    nb = int(cytron.cy_rfil_rela("/cytron/sys/app/sauvegardes_jeu_a_cliquer/", "save.txt"))
+    nb = float(cytron.cy_rfil_rela("/cytron/sys/app/sauvegardes_jeu_a_cliquer/", "save.txt"))
 except:
     cytron.cy_mkfil("/cytron/sys/app/sauvegardes_jeu_a_cliquer/", "save.txt", nb)
